+++ HISTORY.md
@@ -2,6 +2,9 @@
 
 ## in development:
 
+### 2025-10-5
+* blocks and objects: new "object" shape
+
 ### 2025-10-4
 * gui: snapMenu update for JASM